+++ docs/source/conf.py
@@ -22,11 +22,11 @@
 # -- Project information -----------------------------------------------------
 
 project = 'NNMT'
-copyright = '2021, Moritz Layer, Johanna Senk, Simon Essink, Alexander van Meegen, Hannah Bos, Moritz Helias'
+copyright = '2022, Moritz Layer, Johanna Senk, Simon Essink, Alexander van Meegen, Hannah Bos, Moritz Helias'
 author = 'Moritz Layer, Johanna Senk, Simon Essink, Alexander van Meegen, Hannah Bos, Moritz Helias'
 
 # The full version, including alpha/beta/rc tags
-release = '1.0.0'
+release = '1.0.2'
 
 
 # -- General configuration ---------------------------------------------------