+++ nnmt/lif/exp.py
@@ -165,10 +165,10 @@ def _firing_rates(J, K, V_0_rel, V_th_rel, tau_m, tau_r, tau_s, J_ext, K_ext,
     Calculates the stationary firing rate of a neuron with synaptic filter of
     time constant tau_s driven by Gaussian noise with mean mu and standard
     deviation sigma based on :cite:t:`fourcaud2002`, using either a shift of
-    the integration boundaries in the white noise Siegert formula, as derived
-    in :cite:t:`schuecker2015` (default), or a Taylor expansion around
+    the integration boundaries in the white noise Siegert formula, calling
+    :func:`nnmt.lif.exp._firing_rate_shift`, or a Taylor expansion around
     :math:`k = \sqrt{\\tau_\mathrm{s}/\\tau_\mathrm{m}}` of Eq. 4.33 in
-    :cite:t:`fourcaud2002`.
+    :cite:t:`fourcaud2002`, calling :func:`nnmt.lif.exp._firing_rate_taylor`.
 
     Parameters
     ----------
@@ -181,11 +181,11 @@ def _firing_rates(J, K, V_0_rel, V_th_rel, tau_m, tau_r, tau_s, J_ext, K_ext,
     V_th_rel : [float | 1d array]
         Relative threshold potential in V.
     tau_m : [float | 1d array]
-        Membrane time constant in s.
+        Membrane time constant of post-synatic neuron in s.
     tau_r : [float | 1d array]
         Refractory time in s.
-    tau_s : [float | 1d array]
-        Synaptic time constant in s.
+    tau_s : float
+        Pre-synaptic time constant in s.
     J_ext : np.array
         External weight matrix in V.
     K_ext : np.array
@@ -238,10 +238,15 @@ def _firing_rates(J, K, V_0_rel, V_th_rel, tau_m, tau_r, tau_s, J_ext, K_ext,
 @_check_k_in_fast_synaptic_regime
 def _firing_rate_shift(mu, sigma, V_0_rel, V_th_rel, tau_m, tau_r, tau_s):
     """
-    Calculates stationary firing rates including synaptic filtering.
+    Calculates stationary mean firing rates including synaptic filtering.
 
-    Based on :cite:t:`fourcaud2002`, using shift of the integration boundaries
-    in the white noise Siegert formula, as derived in :cite:t:`schuecker2015`.
+    Based on the equation after Eq. 4.33 in :cite:t:`fourcaud2002`, using shift
+    of the integration boundaries in the white noise Siegert formula.
+
+    **Assumptions and approximations**:
+
+    - Diffusion approximation
+    - Fast synapses: :math:`\sqrt{\\tau_\mathrm{s} / \\tau_\mathrm{m}} \ll 1`
 
     Parameters
     ----------
@@ -254,11 +259,11 @@ def _firing_rate_shift(mu, sigma, V_0_rel, V_th_rel, tau_m, tau_r, tau_s):
     V_0_rel : [float | np.array]
         Relative reset potential in V.
     tau_m : [float | 1d array]
-        Membrane time constant in s.
+        Membrane time constant of post-synatic neuron in s.
     tau_r : [float | 1d array]
         Refractory time in s.
-    tau_s : [float | 1d array]
-        Synaptic time constant in s.
+    tau_s : float
+        Pre-synaptic time constant in s.
 
     Returns
     -------
@@ -279,12 +284,17 @@ def _firing_rate_shift(mu, sigma, V_0_rel, V_th_rel, tau_m, tau_r, tau_s):
 @_check_k_in_fast_synaptic_regime
 def _firing_rate_taylor(mu, sigma, V_0_rel, V_th_rel, tau_m, tau_r, tau_s):
     """
-    Calcs stationary firing rates for exp PSCs using a Taylor expansion.
+    Calcs stationary mean firing rates including synaptic filtering.
 
     Calculates the stationary firing rate of a neuron with synaptic filter of
     time constant tau_s driven by Gaussian noise with mean mu and standard
     deviation sigma, using Eq. 4.33 in :cite:t:`fourcaud2002` with Taylor
     expansion around :math:`k = \sqrt{\\tau_\mathrm{s}/\\tau_\mathrm{m}}`.
+
+    **Assumptions and approximations**:
+
+    - Diffusion approximation
+    - Fast synapses: :math:`\sqrt{\\tau_\mathrm{s} / \\tau_\mathrm{m}} \ll 1`
 
     Parameters
     ----------
@@ -297,11 +307,11 @@ def _firing_rate_taylor(mu, sigma, V_0_rel, V_th_rel, tau_m, tau_r, tau_s):
     V_0_rel : [float | np.array]
         Relative reset potential in V.
     tau_m : [float | 1d array]
-        Membrane time constant in s.
+        Membrane time constant of post-synatic neuron in s.
     tau_r : [float | 1d array]
         Refractory time in s.
-    tau_s : [float | 1d array]
-        Synaptic time constant in s.
+    tau_s : float
+        Pre-synaptic time constant in s.
 
     Returns
     -------
@@ -431,7 +441,7 @@ def mean_input(network):
     K : np.array
         Indegree matrix.
     tau_m : [float | 1d array]
-        Membrane time constant in s.
+        Membrane time constant of post-synatic neuron in s.
     J_ext : np.array
         External weight matrix in V.
     K_ext : np.array
@@ -474,7 +484,7 @@ def _mean_input(nu, J, K, tau_m, J_ext, K_ext, nu_ext):
     K : np.array
         In-degree matrix.
     tau_m : [float | 1d array]
-        Membrane time constant in s.
+        Membrane time constant of post-synatic neuron in s.
     J_ext : np.array
         External weight matrix in V.
     K_ext : np.array
@@ -538,7 +548,7 @@ def _std_input(nu, J, K, tau_m, J_ext, K_ext, nu_ext):
     K : np.array
         In-degree matrix.
     tau_m : [float | 1d array]
-        Membrane time constant in s.
+        Membrane time constant of post-synatic neuron in s.
     J_ext : np.array
         External weight matrix in V.
     K_ext : np.array
@@ -632,9 +642,9 @@ def _transfer_function(mu, sigma, tau_m, tau_s, tau_r, V_th_rel, V_0_rel,
     sigma : [float | np.array]
         Standard deviation of neuron activity of one population in V.
     tau_m : [float | np.array]
-        Membrane time constant in s.
-    tau_s : [float | np.array]
-        Synaptic time constant in s.
+        Membrane time constant of post-synatic neuron in s.
+    tau_s : float
+        Pre-synaptic time constant in s.
     tau_r : [float | np.array]
         Refractory time in s.
     V_th_rel : [float | np.array]
@@ -673,15 +683,23 @@ def _transfer_function_shift(mu, sigma, tau_m, tau_s, tau_r, V_th_rel,
     """
     Calcs value of transfer func for one population at given frequency omega.
 
-    Calculates transfer function according to :math:`\tilde{n}` in
-    :cite:t:`schuecker2015`. The expression is to first order equivalent to
-    :func:`nnmt.lif.exp._transfer_function_taylor`. Since the underlying theory
-    is correct to first order, the two expressions are exchangeable.
+    Calculates transfer function based on :math:`\\tilde{n}` in
+    :cite:t:`schuecker2015`. The expression is to first order in
+    :math:`\sqrt{\\tau_\mathrm{s} / \\tau_\mathrm{m}}` equivalent to
+    :func:`nnmt.lif.exp._transfer_function_taylor`.
 
-    The difference here is that the linear response of the system is considered
-    with respect to a perturbation of the input to the current I, leading to an
-    additional low pass filtering 1/(1+i omega tau_s). Compare with the second
-    equation of Eq. 18 and the text below Eq. 29.
+    The difference to the equation in :cite:t:`schuecker2015` is that the
+    linear response of the system is considered with respect to a perturbation
+    of the input to the current I, leading to an additional synaptic low pass
+    filter 1/(1+i omega tau_s). Compare with the second equation of Eq. 18 and
+    the text below Eq. 29.
+
+    **Assumptions and approximations**:
+
+    - Diffusion approximation
+    - Linear response theory
+    - Fast synapses: :math:`\sqrt{\\tau_\mathrm{s} / \\tau_\mathrm{m}} \ll 1`
+    - Low frequencies: :math:`\omega\sqrt{\\tau_\mathrm{m} \\tau_\mathrm{s}} \ll 1`
 
     Parameters
     ----------
@@ -690,9 +708,9 @@ def _transfer_function_shift(mu, sigma, tau_m, tau_s, tau_r, V_th_rel,
     sigma : [float | np.array]
         Standard deviation of neuron activity of one population in V.
     tau_m : [float | np.array]
-        Membrane time constant in s.
-    tau_s : [float | np.array]
-        Synaptic time constant in s.
+        Membrane time constant of post-synatic neuron in s.
+    tau_s : float
+        Pre-synaptic time constant in s.
     tau_r : [float | np.array]
         Refractory time in s.
     V_th_rel : [float | np.array]
@@ -764,8 +782,16 @@ def _transfer_function_taylor(mu, sigma, tau_m, tau_s, tau_r, V_th_rel,
 
     The difference here is that the linear response of the system is considered
     with respect to a perturbation of the input to the current I, leading to an
-    additional low pass filtering 1/(1+i omega tau_s).
-    Compare with the second equation of Eq. 18 and the text below Eq. 29.
+    additional synaptic low pass filter 1/(1+i omega tau_s). Compare with the
+    second equation of Eq. 18 and the text below Eq. 29.
+
+    **Assumptions and approximations**:
+
+    - Diffusion approximation
+    - Linear response theory
+    - Fast synapses: :math:`\sqrt{\\tau_\mathrm{s} / \\tau_\mathrm{m}} \ll 1`
+    - Low frequencies: :math:`\omega\sqrt{\\tau_\mathrm{m} \\tau_\mathrm{s}}
+      \ll 1`
 
     Parameters
     ----------
@@ -774,9 +800,9 @@ def _transfer_function_taylor(mu, sigma, tau_m, tau_s, tau_r, V_th_rel,
     sigma : [float | np.array]
         Standard deviation of neuron activity of one population in V.
     tau_m : [float | np.array]
-        Membrane time constant in s.
-    tau_s : [float | np.array]
-        Synaptic time constant in s.
+        Membrane time constant of post-synatic neuron in s.
+    tau_s : float
+        Pre-synaptic time constant in s.
     tau_r : [float | np.array]
         Refractory time in s.
     V_th_rel : [float | np.array]
@@ -899,7 +925,7 @@ def _fit_transfer_function(transfer_function, omegas, tau_m, J, K):
     Fits the transfer function (tf) of a low-pass filter to the passed tf.
 
     For details of the fitting procedure see
-    :func:`nnmt._general._fit_transfer_function`.
+    :func:`nnmt.lif._general._fit_transfer_function`.
 
     For details of the theory refer to
     :cite:t:`senk2020`, Sec. F 'Comparison of neural-field and spiking models'.
@@ -912,7 +938,7 @@ def _fit_transfer_function(transfer_function, omegas, tau_m, J, K):
     omegas : [float | np.ndarray]
         Input frequencies to population in Hz.
     tau_m : float
-        Membrane time constant in s.
+        Membrane time constant of post-synatic neuron in s.
     J : np.array
         Weight matrix in V.
     K : np.array
@@ -975,6 +1001,11 @@ def _derivative_of_firing_rates_wrt_mean_input(mu, sigma, V_0_rel, V_th_rel,
 
     See Appendix B in :cite:t:`schuecker2014`.
 
+    **Assumptions and approximations**:
+
+    - Diffusion approximation
+    - Fast synapses: :math:`\sqrt{\\tau_\mathrm{s} / \\tau_\mathrm{m}} \ll 1`
+
     Parameters
     ----------
     mu : [float | np.ndarray]
@@ -986,11 +1017,11 @@ def _derivative_of_firing_rates_wrt_mean_input(mu, sigma, V_0_rel, V_th_rel,
     V_0_rel : [float | np.ndarray]
         Relative reset potential in V.
     tau_m : [float | np.ndarray]
-        Membrane time constant in s.
+        Membrane time constant of post-synatic neuron in s.
     tau_r : [float | np.ndarray]
         Refractory time in s.
     tau_s : [float | np.ndarray]
-        Synaptic time constant in s.
+        Pre-synaptic time constant in s.
 
     Returns
     -------
@@ -1027,6 +1058,11 @@ def _Phi(s):
 def _Psi(z, x):
     """
     Calcs Psi(z,x)=exp(x**2/4)*U(z,x), with U(z,x) the parabolic cylinder func.
+
+    The mpmath.pcfu() is equivalent to Eq. 19.12.3 in:cite:t:`Abramowitz74`
+    with U(a,-x). The arguments (a, z) of mpmath.pcfu() used in the
+    documentation https://mpmath.org/doc/current/functions/bessel.html?highlight=pcfu#mpmath.pcfu
+    are renamed to (z, x) here.
     """
     parabolic_cylinder_fn = pcfu_vec(z, -x).astype(complex)
     return np.exp(0.25 * x**2) * parabolic_cylinder_fn
@@ -1073,6 +1109,11 @@ def _derivative_of_firing_rates_wrt_input_rate(
 
     See Eq. A.3 in Appendix A of :cite:t:`helias2013`.
 
+    **Assumptions and approximations**:
+
+    - Diffusion approximation
+    - Fast synapses: :math:`\sqrt{\\tau_\mathrm{s} / \\tau_\mathrm{m}} \ll 1`
+
     Parameters
     ----------
     mu : [float | np.ndarray]
@@ -1080,9 +1121,9 @@ def _derivative_of_firing_rates_wrt_input_rate(
     sigma :
         Standard deviation of neuron activity in V.
     tau_m : [float | np.ndarray]
-        Membrane time constant in s.
-    tau_s : [float | np.ndarray]
-        Synaptic time constant in s.
+        Membrane time constant of post-synatic neuron in s.
+    tau_s : float
+        Pre-synaptic time constant in s.
     tau_r : [float | np.ndarray]
         Refractory time in s.
     V_th_rel : [float | np.ndarray]
@@ -1196,7 +1237,7 @@ def _effective_connectivity(transfer_function, D, J, K, tau_m):
     K : np.ndarray
         Indegree matrix.
     tau_m : float
-        Membrane time constant in s.
+        Membrane time constant of post-synatic neuron in s.
 
     Returns
     -------
@@ -1435,6 +1476,10 @@ def sensitivity_measure(network, frequency,
             Index of critical_frequency in all analysis frequencies.
         critical_eigenvalue : np.complex
             Critical eigenvalue.
+        left_eigenvector : np.ndarray
+            Left eigenvector corresponding to the critical eigenvalue.
+        right_eigenvector : np.ndarray
+            Right eigenvector corresponding to the critical eigenvalue.
         k : np.complex
             Vector point from critical eigenvalue to complex(1,0).
         k_per : np.complex
@@ -1510,6 +1555,10 @@ def _sensitivity_measure(effective_connectivity, frequency,
             Index of critical_frequency in all analysis frequencies.
         critical_eigenvalue : np.complex
             Critical eigenvalue.
+        left_eigenvector : np.ndarray
+            Left eigenvector corresponding to the critical eigenvalue.
+        right_eigenvector : np.ndarray
+            Right eigenvector corresponding to the critical eigenvalue.
         k : np.complex
             Vector point from critical eigenvalue to complex(1,0).
         k_per : np.complex
@@ -1586,6 +1635,8 @@ def _sensitivity_measure(effective_connectivity, frequency,
             'critical_frequency': frequency,
             'critical_frequency_index': frequency_index,
             'critical_eigenvalue': critical_eigenvalue,
+            'left_eigenvector': U_l[:, eigenvalue_index],
+            'right_eigenvector': U_r[:, eigenvalue_index],
             'k': k,
             'k_per': k_per,
             'sensitivity': T,
@@ -1744,7 +1795,7 @@ def _power_spectra(nu, effective_connectivity, J, K, N, tau_m):
     N : np.ndarray
         Number of neurons in each population.
     tau_m : [float | np.narray]
-        Membrane time constant in s.
+        Membrane time constant of post-synatic neuron in s.
 
     Returns
     -------
@@ -1837,11 +1888,11 @@ def _external_rates_for_fixed_input(mu_set, sigma_set,
     V_0_rel : [float | np.array]
         Relative reset potential in V.
     tau_m : [float | 1d array]
-        Membrane time constant in s.
+        Membrane time constant of post-synatic neuron in s.
     tau_r : [float | 1d array]
         Refractory time in s.
-    tau_s : [float | 1d array]
-        Synaptic time constant in s.
+    tau_s : float
+        Pre-synaptic time constant in s.
     J_ext : np.array
         External weight matrix in V.
     K_ext : np.array